Moved reducers and actions into new folders and can now list out what we enter in the input
