--- src/actionCreators.js
@@ -1,5 +1,0 @@
-import { SET_ITEMS } from './actions'
-
-export function setItems(item) {
-  return { type: SET_ITEMS, payload: item }
-}
