--- src/actions.js
@@ -1,2 +1,0 @@
-export const SET_ITEMS = 'SET_ITEMS'
-
