+++ src/actions/index.js
@@ -1,0 +1,4 @@
+export const setItems = item => ({
+  type: 'SET_ITEMS',
+  payload: item
+})
