+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import literal from './literal'
+
+export default combineReducers({
+  literal
+})
