+++ src/store.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux'
-import reducer from './reducers'
+import reducer from './reducers/literal'
 
 const store = createStore(reducer)
